+++ apps/showcase/src/pages/ComponentDetailPage.vue
@@ -158,6 +158,7 @@ const previewRegistry = {
   'ds-segmented-states': defineAsyncComponent(() => import('../demos/components/ds-segmented/GrSegmentedStatesDemo.vue')),
   'ds-select-builder': defineAsyncComponent(() => import('../demos/components/ds-select/GrSelectBuilderDemo.vue')),
   'ds-select-custom-value': defineAsyncComponent(() => import('../demos/components/ds-select/GrSelectCustomValueDemo.vue')),
+  'ds-select-groups': defineAsyncComponent(() => import('../demos/components/ds-select/GrSelectGroupsDemo.vue')),
   'ds-select-native-modes': defineAsyncComponent(() => import('../demos/components/ds-select/GrSelectModesDemo.vue')),
   'ds-select-panel-multiple': defineAsyncComponent(() => import('../demos/components/ds-select/GrSelectPanelDemo.vue')),
   'ds-sidebar-basic-sections': defineAsyncComponent(() => import('../demos/components/ds-sidebar/GrSidebarBasicSectionsDemo.vue')),